factored web-related functions out of conn2lines script and into separate web module so I can write other scripts that use them
