+++ conn2lines.py
@@ -13,17 +13,13 @@ Crawl graphs of typed Pleiades connections and build straight lines between them
 """
 
 from airtight.cli import configure_commandline
-from datetime import timedelta
 import logging
 from pprint import pformat, pprint
 from rdflib import Graph, Namespace, URIRef
-from urllib.parse import urlparse
-from validators import url as valid_uri
-from webiquette.webi import Webi
+from web import *
 
 logger = logging.getLogger(__name__)
 
-CACHE_EXPIRATION = timedelta(hours=24)
 DEFAULT_LOG_LEVEL = logging.WARNING
 OPTIONAL_ARGUMENTS = [
     [
@@ -54,88 +50,12 @@ POSITIONAL_ARGUMENTS = [
     # each row is a list with 3 elements: name, type, help
     ["start_id", str, "Pleiades URI to start with"],
 ]
-BASE_URI = "https://pleiades.stoa.org/places/"
 ns_dcterms = Namespace("http://purl.org/dc/terms/")
 ns_geo = Namespace("http://www.w3.org/2003/01/geo/wgs84_pos#")
 ns_pleiades_relationship_types = Namespace(
     "https://pleiades.stoa.org/vocabularies/relationship-types/"
 )
 ns_pleiades_places = Namespace("https://pleiades.stoa.org/places/vocab#")
-
-
-def get_ttl(webi, puri) -> str:
-    ### download and return RDF/TTL for puri as text
-    r = webi.get(puri + "/turtle")
-    if r.status_code != 200:
-        r.raise_for_status()
-    return r.text
-
-
-def get_json(webi, puri) -> dict:
-    ### download and return JSON for puri as a python dict
-    r = webi.get(puri + "/json")
-    if r.status_code != 200:
-        r.raise_for_status()
-    return r.json()
-
-
-def get_place(webi, puri) -> Graph:
-    ### download and parse data about the place at puri and return it as an RDFLIB Graph
-    ttl = get_ttl(webi, puri)
-    g = Graph()
-    g.parse(data=ttl, format="turtle")
-    logger.debug(f"RDF for {puri} has {len(g)} triples.")
-
-    # annoyingly, pleiades RDF doesn't have the connections from the database, so we have to build and add them from the JSON
-    j = get_json(webi, puri)
-    connections = j["connections"]
-    logger.debug(f"JSON for {puri} has {len(connections)} connections.")
-    for c in connections:
-        g.add((URIRef(puri), URIRef(c["connectionTypeURI"]), URIRef(c["connectsTo"])))
-    logger.debug(f"Modified RDF for {puri} now has {len(g)} triples.")
-    return g
-
-
-def get_repr_point(webi, puri) -> tuple:
-    ### get the representative point for a puri from its json
-    j = get_json(webi, puri)
-    try:
-        return tuple(j["reprPoint"])
-    except TypeError:
-        return None
-
-
-def get_web_interface() -> Webi:
-    ### instantiate and return a "webi" web interface object for the Pleiades website
-    headers = {
-        "User-Agent": "Pleiades4Sebs2023/0.1",
-        "from": "pleiades.admin@nyu.edu",
-    }
-    return Webi(
-        netloc=urlparse(BASE_URI).netloc, headers=headers, expire_after=CACHE_EXPIRATION
-    )
-
-
-def raise_invalid_uri(v):
-    ### raise a value error for an invalid Pleiades URI
-    raise ValueError(f"Invalid Pleiades URI or ID: {v}")
-
-
-def validate_id(v):
-    ### validate and coerce if necessary a Pleiades place ID or URI
-    id = v
-    while True:
-        if not (valid_uri(id)):
-            try:
-                int(id)
-            except ValueError:
-                raise_invalid_uri(v)
-            else:
-                id = BASE_URI + id
-        elif not id.startswith(BASE_URI):
-            raise_invalid_uri(v)
-        else:
-            return id
 
 
 def main(**kwargs):
@@ -156,7 +76,7 @@ def main(**kwargs):
         if not puris_to_parse:
             break
         for puri in puris_to_parse:
-            connected_places[puri] = get_place(webi, puri)
+            connected_places[puri] = get_place_graph(webi, puri)
             for contype in conntypes:
                 for s, p, o in connected_places[puri].triples((None, contype, None)):
                     try:
